+++ template/app/src/client/admin/components/CheckboxOne.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { cn } from '../../../shared/utils';
+import { cn } from '../../cn';
 
 const CheckboxOne = () => {
   const [isChecked, setIsChecked] = useState<boolean>(false);
