+++ template/app/src/client/admin/components/Header.tsx
@@ -2,7 +2,7 @@ import { type AuthUser } from 'wasp/auth/types';
 import DarkModeSwitcher from './DarkModeSwitcher';
 import MessageButton from './MessageButton';
 import DropdownUser from '../../components/DropdownUser';
-import { cn } from '../../../shared/utils';
+import { cn } from '../../cn';
 
 const Header = (props: {
   sidebarOpen: string | boolean | undefined;
