+++ template/app/src/client/landing-page/urls.ts
@@ -1,0 +1,2 @@
+export const DocsUrl = 'https://docs.opensaas.sh';
+export const BlogUrl = 'https://docs.opensaas.sh/blog';
