--- template/app/src/shared/constants.ts
@@ -1,2 +1,0 @@
-export const DOCS_URL = 'https://docs.opensaas.sh';
-export const BLOG_URL = 'https://docs.opensaas.sh/blog';
